feat: add number formatter and add sold numbers
